Rename simple to standard
