+++ common/persistence/visibility/store/standard/converter_test.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package simple
+package standard
 
 import (
 	"fmt"
